# openapi/top.yaml
swagger: "2.0"
...
info:
  title: Rosbag Bazaar API
  description: API to access the Rosbag Bazaar service
  version: "0.1.1"
  contact:
    email: hhendrik@student.ethz.ch
basePath: /api/v0
schemes:
  - https
  - http
...
securityDefinitions:
  basicAuth:
    type: basic
security:
 - basicAuth: []

###########################
# API Paths
###########################
paths:
...
  /queue:
    get:
      tags:
        - basic
      summary: List task queue
      operationId: list_queue
      parameters:
        # Pagination
        - name: limit
          in: query
          required: false
          type: integer
        - name: offset
          in: query
          required: false
          type: integer

        # Ordering
        - name: ordering
          in: query
          required: false
          type: string

        # Filtering
        - name: running
          description: Include running tasks, default is true
          in: query
          required: false
          type: string
        - name: finished
          description: Include finished tasks, default is true
          in: query
          required: false
          type: string
        - name: queued
          description: Include queued tasks, default is true
          in: query
          required: false
          type: string
      responses:
        '200':
          description: Returns an array of tasks.
          schema:
            type: array
            items:
              $ref: '#/definitions/TaskSummary'
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
    post:
      tags:
        - basic
      summary: Create a new task
      operationId: new_task
      parameters:
        - name: task
          in: body
          description: The task
          required: true
          schema:
            $ref: '#/definitions/TaskDetailed'
      responses:
        '200':
          description: The new task
          schema:
            $ref: "#/definitions/TaskDetailed"
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
  /queue/dequeue:
    post:
      tags:
        - basic
      summary: Take a task from the queue
      operationId: dequeue_task
      parameters:
        - name: worker_name
          in: query
          description: Name of the worker trying to acquire a task
          required: true
          type: string
        - name: tasks
          in: query
          description: Tasks that the worker can do (any or a list of tasks)
          required: true
          type: string
        - name: labels
          in: query
          description: Labels the worker wants to do
          required: true
          type: string
      responses:
        '200':
          description: The task
          schema:
            $ref: "#/definitions/TaskDetailed"
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
  /queue/{task_identifier}:
    get:
      tags:
        - basic
      summary: Take a task from the queue
      operationId: get_task
      parameters:
        - name: task_identifier
          in: path
          required: true
          type: string
      responses:
        '200':
          description: The task
          schema:
            $ref: "#/definitions/TaskDetailed"
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
    put:
      tags:
        - basic
      summary: Update a task
      operationId: put_task
      parameters:
        - name: task_identifier
          in: path
          required: true
          type: string
        - name: task
          in: body
          description: The task
          required: true
          schema:
            $ref: '#/definitions/TaskDetailed'
      responses:
        '200':
          description: The updated task
          schema:
            $ref: "#/definitions/TaskDetailed"
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
    patch:
      tags:
        - basic
      summary: Partial update of task (this only supports a few fields)
      operationId: patch_task
      parameters:
        - name: task_identifier
          in: path
          required: true
          type: string
        - name: task
          in: body
          description: Fields to update
          required: true
          schema:
           type: object
      responses:
        '200':
          description: Returns the task information
          schema:
            $ref: "#/definitions/TaskDetailed"
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
    post:
      tags:
        - basic
      summary: Perform an action on the task
      operationId: do_task_action
      parameters:
        - name: task_identifier
          in: path
          required: true
          type: string
        - name: action
          in: query
          description: Action to perform (cancel/prio_up)
          required: true
          type: string
        - name: task
          in: body
          description: The task, required depending on the action
          schema:
            $ref: '#/definitions/TaskDetailed'
      responses:
        '200':
          description: The updated task
          schema:
            $ref: "#/definitions/TaskDetailed"
        '400':
          description: General error
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "rbb_server.controllers.queue_controller"
...
definitions:
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
...
  TaskSummary:
    type: object
    discriminator: detail_type
    properties:
      detail_type:
        type: string
        example: 'TaskSummary'
      identifier:
        type: string
        example: '1'
      priority:
        type: number
        example: 1000
      description:
        type: string
        description: Simple description to identify the type of task
      assigned_to:
        type: string
        description: Name of worker assigned to the task.
      created:
        type: string
        format: 'date-time'
        description: Date the task was added to the queue.
      last_updated:
        type: string
        format: 'date-time'
        description: Date the task was updated to the queue.
      state:
        type: integer
        description: State of the task.
      task:
        type: string
        description: The task.
      success:
        type: boolean
      runtime:
        type: number
      worker_labels:
        type: string
        description: Worker that can take this task.
    required:
      - detail_type
      - identifier
      - priority
      - description
      - assigned_to
      - created
      - state
      - task
      - success
      - runtime
      - worker_labels
      - last_updated
  TaskDetailed:
    allOf:
      - $ref: '#/definitions/TaskSummary'
      - type: object
        properties:
          config:
            type: object
            description: Configuration of the task.
          result:
            type: object
            description: Result of the task
          log:
            type: string
            description: Standard output of the task
        required:
          - topics
          - products
          - comment
...